reshape dirs and add threads & checker
