+++ Backend/twitter/api/serializer.py
@@ -1,4 +1,4 @@
-from .models import (
+from ..models import (
     TwitterAccount,
     Agent,
     Project,
