--- Backend/twitter/tasks.py
@@ -1,10 +1,0 @@
-from celery import shared_task 
-
-
-
-
-
-@shared_task
-def loggering():
-    print(f"\nhhhhhhhhh\n")
-    return
